+++ FuzzySharp/Edits/MatchingBlock.cs
@@ -1,11 +1,10 @@
-﻿namespace Raffinert.FuzzySharp.Edits
-{
-    public class MatchingBlock
-    {
-        public int SourcePos { get; set; }
-        public int DestPos { get; set; }
-        public int Length { get; set; }
+﻿namespace Raffinert.FuzzySharp.Edits;
 
-        public override string ToString() => $"({SourcePos},{DestPos},{Length})";
-    }
+public record MatchingBlock
+{
+    public int SourcePos { get; set; }
+    public int DestPos { get; set; }
+    public int Length { get; set; }
+
+    public override string ToString() => $"({SourcePos},{DestPos},{Length})";
 }
