+++ FuzzySharp/SimilarityRatio/Scorer/Generic/IRatioScorer.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
+
+public interface IRatioScorer<in T> where T : IEquatable<T>
 {
-    public interface IRatioScorer<in T> where T : IEquatable<T>
-    {
-        int Score(T[] input1, T[] input2);
-    }
+    int Score(T[] input1, T[] input2);
 }
