+++ FuzzySharp/SimilarityRatio/Scorer/IRatioScorer.cs
@@ -1,10 +1,9 @@
 ﻿using Raffinert.FuzzySharp.PreProcess;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer;
+
+public interface IRatioScorer
 {
-    public interface IRatioScorer
-    {
-        int Score(string input1, string input2);
-        int Score(string input1, string input2, PreprocessMode preprocessMode);
-    }
+    int Score(string input1, string input2);
+    int Score(string input1, string input2, PreprocessMode preprocessMode);
 }
