+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Generic/StrategySensitiveScorerBase.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive.Generic
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive.Generic;
+
+public abstract class StrategySensitiveScorerBase<T> : ScorerBase<T> where T : IEquatable<T>
 {
-    public abstract class StrategySensitiveScorerBase<T> : ScorerBase<T> where T : IEquatable<T>
-    {
-        protected abstract Func<T[], T[], int> Scorer { get; }
-    }
+    protected abstract Func<T[], T[], int> Scorer { get; }
 }
