+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Simple/DefaultRatioScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class DefaultRatioScorer : SimpleRatioScorerBase
 {
-    public class DefaultRatioScorer : SimpleRatioScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
 }
