+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Simple/PartialRatioScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class PartialRatioScorer : SimpleRatioScorerBase
 {
-    public class PartialRatioScorer : SimpleRatioScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
 }
