+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/StrategySensitiveScorerBase.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public abstract class StrategySensitiveScorerBase : ScorerBase
 {
-    public abstract class StrategySensitiveScorerBase : ScorerBase
-    {
-        protected abstract FuzzySharp.Scorer Scorer { get; }
-    }
+    protected abstract FuzzySharp.Scorer Scorer { get; }
 }
