+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenAbbreviation/PartialTokenAbbreviationScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class PartialTokenAbbreviationScorer : TokenAbbreviationScorerBase
 {
-    public class PartialTokenAbbreviationScorer : TokenAbbreviationScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
 }
