+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenDifference/PartialTokenDifferenceScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy.Generic;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class PartialTokenDifferenceScorer : TokenDifferenceScorerBase
 {
-    public class PartialTokenDifferenceScorer : TokenDifferenceScorerBase
-    {
-        protected override Func<string[], string[], int> Scorer => PartialRatioStrategy<string>.Calculate;
-    }
+    protected override Func<string[], string[], int> Scorer => PartialRatioStrategy<string>.Calculate;
 }
