+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenDifference/TokenDifferenceScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy.Generic;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class TokenDifferenceScorer : TokenDifferenceScorerBase
 {
-    public class TokenDifferenceScorer : TokenDifferenceScorerBase
-    {
-        protected override Func<string[], string[], int> Scorer => DefaultRatioStrategy<string>.Calculate;
-    }
+    protected override Func<string[], string[], int> Scorer => DefaultRatioStrategy<string>.Calculate;
 }
