+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenInitialism/PartialTokenInitialismScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class PartialTokenInitialismScorer : TokenInitialismScorerBase
 {
-    public class PartialTokenInitialismScorer : TokenInitialismScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
 }
