+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenInitialism/TokenInitialismScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class TokenInitialismScorer : TokenInitialismScorerBase
 {
-    public class TokenInitialismScorer : TokenInitialismScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
 }
