+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSet/PartialTokenSetScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class PartialTokenSetScorer : TokenSetScorerBase
 {
-    public class PartialTokenSetScorer : TokenSetScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
 }
