+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSet/TokenSetScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class TokenSetScorer : TokenSetScorerBase
 {
-    public class TokenSetScorer : TokenSetScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
 }
