+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSort/PartialTokenSortScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class PartialTokenSortScorer : TokenSortScorerBase
 {
-    public class PartialTokenSortScorer : TokenSortScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
 }
