+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSort/TokenSortAlgorithm.cs
@@ -1,15 +1,14 @@
 ﻿using Raffinert.FuzzySharp.Extensions;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
-{
-    public abstract class TokenSortScorerBase : StrategySensitiveScorerBase
-    {
-        public override int Score(string input1, string input2)
-        {
-            var sorted1 = input1.NormalizeSpacesAndSort();
-            var sorted2 = input2.NormalizeSpacesAndSort();
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
 
-            return Scorer(sorted1, sorted2);
-        }
+public abstract class TokenSortScorerBase : StrategySensitiveScorerBase
+{
+    public override int Score(string input1, string input2)
+    {
+        var sorted1 = input1.NormalizeSpacesAndSort();
+        var sorted2 = input2.NormalizeSpacesAndSort();
+
+        return Scorer(sorted1, sorted2);
     }
 }
