+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSort/TokenSortScorer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public class TokenSortScorer : TokenSortScorerBase
 {
-    public class TokenSortScorer : TokenSortScorerBase
-    {
-        protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
-    }
+    protected override FuzzySharp.Scorer Scorer => DefaultRatioStrategy.Calculate;
 }
