+++ FuzzySharp/SimilarityRatio/ScorerCache.cs
@@ -1,17 +1,16 @@
 ﻿using System.Runtime.CompilerServices;
 using Raffinert.FuzzySharp.SimilarityRatio.Scorer;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio
-{
-    public static class ScorerCache
-    {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static IRatioScorer Get<T>() where T : IRatioScorer, new() => GenericCache<T>.Instance;
+namespace Raffinert.FuzzySharp.SimilarityRatio;
 
-        private static class GenericCache<T>
-            where T : IRatioScorer, new()
-        {
-            public static readonly T Instance = new T();
-        }
+public static class ScorerCache
+{
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static IRatioScorer Get<T>() where T : IRatioScorer, new() => GenericCache<T>.Instance;
+
+    private static class GenericCache<T>
+        where T : IRatioScorer, new()
+    {
+        public static readonly T Instance = new T();
     }
 }
