+++ FuzzySharp/SimilarityRatio/Strategy/DefaultRatioStrategy.cs
@@ -1,16 +1,15 @@
 ﻿using System;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Strategy
+namespace Raffinert.FuzzySharp.SimilarityRatio.Strategy;
+
+internal static class DefaultRatioStrategy
 {
-    internal static class DefaultRatioStrategy
+    public static int Calculate(string input1, string input2)
     {
-        public static int Calculate(string input1, string input2)
+        if (input1.Length == 0 || input2.Length == 0)
         {
-            if (input1.Length == 0 || input2.Length == 0)
-            {
-                return 0;
-            }
-            return (int)Math.Round(100 * Indel.NormalizedSimilarity(input1.AsSpan(), input2.AsSpan()));
+            return 0;
         }
+        return (int)Math.Round(100 * Indel.NormalizedSimilarity(input1.AsSpan(), input2.AsSpan()));
     }
 }
