+++ FuzzySharp/SimilarityRatio/Strategy/Generic/PartialRatioStrategyT.cs
@@ -1,59 +1,58 @@
 ﻿using System;
 
-namespace Raffinert.FuzzySharp.SimilarityRatio.Strategy.Generic
+namespace Raffinert.FuzzySharp.SimilarityRatio.Strategy.Generic;
+
+internal static class PartialRatioStrategy<T> where T : IEquatable<T>
 {
-    internal static class PartialRatioStrategy<T> where T : IEquatable<T>
+    public static int Calculate(T[] input1, T[] input2)
     {
-        public static int Calculate(T[] input1, T[] input2)
+        T[] shorter;
+        T[] longer;
+
+        if (input1.Length == 0 || input2.Length == 0)
         {
-            T[] shorter;
-            T[] longer;
-
-            if (input1.Length == 0 || input2.Length == 0)
-            {
-                return 0;
-            }
-
-            if (input1.Length < input2.Length)
-            {
-                shorter = input1;
-                longer  = input2;
-            }
-            else
-            {
-                shorter = input2;
-                longer  = input1;
-            }
-
-            var matchingBlocks = Levenshtein.GetMatchingBlocks(shorter, longer);
-
-            double maxScore = 0;
-
-            foreach (var matchingBlock in matchingBlocks)
-            {
-                int dist = matchingBlock.DestPos - matchingBlock.SourcePos;
-
-                int longStart = dist > 0 ? dist : 0;
-                int longEnd   = longStart + shorter.Length;
-
-                if (longEnd > longer.Length) longEnd = longer.Length;
-
-                var longSubstr = longer.AsSpan()[longStart..longEnd];
-
-                double ratio = Indel.NormalizedSimilarity(new ReadOnlySpan<T>(shorter), longSubstr);
-
-                if (ratio > .995)
-                {
-                    return 100;
-                }
-
-                if (ratio > maxScore)
-                {
-                    maxScore = ratio;
-                }
-            }
-
-            return (int)Math.Round(100 * maxScore);
+            return 0;
         }
+
+        if (input1.Length < input2.Length)
+        {
+            shorter = input1;
+            longer  = input2;
+        }
+        else
+        {
+            shorter = input2;
+            longer  = input1;
+        }
+
+        var matchingBlocks = Levenshtein.GetMatchingBlocks(shorter, longer);
+
+        double maxScore = 0;
+
+        foreach (var matchingBlock in matchingBlocks)
+        {
+            int dist = matchingBlock.DestPos - matchingBlock.SourcePos;
+
+            int longStart = dist > 0 ? dist : 0;
+            int longEnd   = longStart + shorter.Length;
+
+            if (longEnd > longer.Length) longEnd = longer.Length;
+
+            var longSubstr = (ReadOnlySpan<T>)longer.AsSpan()[longStart..longEnd];
+
+            double ratio = Indel.NormalizedSimilarity(new ReadOnlySpan<T>(shorter), longSubstr);
+
+            if (ratio > .995)
+            {
+                return 100;
+            }
+
+            if (ratio > maxScore)
+            {
+                maxScore = ratio;
+            }
+        }
+
+        return (int)Math.Round(100 * maxScore);
     }
 }
